move sockets to common/network
