+++ src/common/network/server_socket.hpp
@@ -3,7 +3,7 @@
 #include <memory>
 
 #include "socket.hpp"
-#include "autos/auto_fd.hpp"
+#include "../autos/auto_fd.hpp"
 
 class ServerSocket final
 {
