+++ src/common/network/socket.cpp
@@ -5,8 +5,8 @@
 
 #include "server_socket.hpp"
 #include "socket.hpp"
-#include "os_utils.hpp"
-#include "unix.hpp"
+#include "../os_utils.hpp"
+#include "../unix.hpp"
 
 Socket::Socket(const int socket_fd)
     : _socket_fd(socket_fd)
