+++ src/netlink.hpp
@@ -1,5 +1,5 @@
 #include <memory>
-#include "common/socket.hpp"
+#include "common/network/socket.hpp"
 
 namespace netlink
 {
